+++ tests/it_metar.rs
@@ -113,3 +113,8 @@ fn it_metar_wind_shear() -> Result<()> {
 fn it_metar_sea() -> Result<()> {
     it_metar_template("it_sea_input.txt", "it_sea_output.json", "noaa-metar-cycles")
 }
+
+#[test]
+fn it_metar_trend() -> Result<()> {
+    it_metar_template("it_trend_input.txt", "it_trend_output.json", "noaa-metar-cycles")
+}